+++ packages/utils/src/fallbackSubgraph.ts
@@ -1,6 +1,9 @@
-import { getHostnameFromRegex } from ".";
+import { request } from "graphql-request";
+
 import { NxtpError } from "./error";
 import { getSubgraphHealth } from "./subgraphHealth";
+
+import { getHostnameFromRegex } from ".";
 
 export type SubgraphSyncRecord = {
   uri: string;
@@ -21,6 +24,7 @@ type CallMetric = {
 };
 
 type Subgraph<T extends SubgraphSdk> = {
+  url: string;
   client: T;
   record: SubgraphSyncRecord;
   priority: number;
@@ -73,21 +77,22 @@ export class FallbackSubgraph<T extends SubgraphSdk> {
   constructor(
     private readonly chainId: number,
     // We use the URIs in sync records for reference in logging.
-    sdks: { client: T; uri: string }[],
+    sdks: { client: T; url: string }[],
     private readonly maxLag: number,
     private readonly stallTimeout = 10_000,
   ) {
-    const getSubgraphName = (uri: string) => {
-      const split = uri.split("/");
+    const getSubgraphName = (url: string) => {
+      const split = url.split("/");
       return split[split.length - 1];
     };
-    this.subgraphs = sdks.map(({ client, uri }) => ({
+    this.subgraphs = sdks.map(({ client, url }) => ({
+      url,
       client,
       record: {
         // Typically used for logging, distinguishing between which subgraph is which, so we can monitor
         // which ones are most in sync.
-        uri: getHostnameFromRegex(uri) ? getHostnameFromRegex(uri)!.split(".").slice(0, -1).join(".") : uri,
-        name: getSubgraphName(uri),
+        uri: getHostnameFromRegex(url) ? getHostnameFromRegex(url)!.split(".").slice(0, -1).join(".") : url,
+        name: getSubgraphName(url),
         synced: true,
         latestBlock: -1,
         syncedBlock: -1,
@@ -105,6 +110,19 @@ export class FallbackSubgraph<T extends SubgraphSdk> {
   }
 
   /**
+   * Raw string query method.
+   *
+   * @param query - GraphQL query string.
+   *
+   * @returns any, whatever the expected GraphQL response is.
+   */
+  public async query(query: string): Promise<any> {
+    this.request((_, url) => {
+      return request(url, query);
+    });
+  }
+
+  /**
    * Make an SDK request using the fallback subgraph wrapper.
    *
    * @param method - anonymous callback function that takes an SdkLike client and executes a subgraph sdk method.
@@ -114,7 +132,7 @@ export class FallbackSubgraph<T extends SubgraphSdk> {
    * @returns A Promise of the generic type.
    * @throws Error if the subgraphs are out of sync (and syncRequired is true).
    */
-  public async request<Q>(method: (client: T) => Promise<Q>, syncRequired = false): Promise<Q> {
+  public async request<Q>(method: (client: T, url: string) => Promise<Q>, syncRequired = false): Promise<Q> {
     // If subgraph sync is requied, we'll check that all subgraphs are in sync before making the request.
     if (syncRequired && !this.inSync) {
       throw new Error(`All subgraphs out of sync on chain ${this.chainId}; unable to handle request.`);
@@ -129,7 +147,7 @@ export class FallbackSubgraph<T extends SubgraphSdk> {
             const startTime = Date.now();
             let success = false;
             try {
-              const result = await method(subgraph.client);
+              const result = await method(subgraph.client, subgraph.url);
               success = true;
               resolve(result);
             } catch (e) {
@@ -176,7 +194,7 @@ export class FallbackSubgraph<T extends SubgraphSdk> {
       }
     };
 
-    // Using a Promise.all here to ensure we do our GetBlockNumber queries in parallel.
+    // Using a Promise.all here to ensure we do all queries/requests in parallel.
     await Promise.all(
       this.subgraphs.map(async (subgraph, index) => {
         const errors: any[] = [];