+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - [txservice] error types
 - [utils] subgraph health query method implementation
+- [sdk] Add generic subgraph query method
+- [sdk] Add generic subgraph query method
 
 # 0.0.101